fix saving new file

+++ 8beatMap/Form1.cs
@@ -598,8 +598,15 @@ namespace _8beatMap
         private void SaveChartBtn_Click(object sender, EventArgs e)
         {
             if (saveFileDialog1.ShowDialog() == DialogResult.OK)
-                System.IO.File.WriteAllText(saveFileDialog1.FileName + ".tmp", Notedata.ConvertChartToJson_Small(chart));
-            System.IO.File.Replace(saveFileDialog1.FileName + ".tmp", saveFileDialog1.FileName, saveFileDialog1.FileName + ".bak");
+                if (System.IO.File.Exists(saveFileDialog1.FileName))
+                {
+                    System.IO.File.WriteAllText(saveFileDialog1.FileName + ".tmp", Notedata.ConvertChartToJson_Small(chart));
+                    System.IO.File.Replace(saveFileDialog1.FileName + ".tmp", saveFileDialog1.FileName, saveFileDialog1.FileName + ".bak");
+                }
+                else
+                {
+                    System.IO.File.WriteAllText(saveFileDialog1.FileName, Notedata.ConvertChartToJson_Small(chart));
+                }
         }
 
         private void OpenMusicButton_Click(object sender, EventArgs e)
